First stage of refactoring towards a service layer
