+++ pinax/stripe/management/commands/init_customers.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 
 from django.contrib.auth import get_user_model
 
-from ...models import Customer
+from ...actions import CustomerProxy
 
 
 class Command(BaseCommand):
@@ -12,5 +12,5 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         User = get_user_model()
         for user in User.objects.filter(customer__isnull=True):
-            Customer.create(user=user)
+            CustomerProxy.create(user=user)
             print("Created customer for {0}".format(user.email))
