+++ pinax/stripe/middleware.py
@@ -1,8 +1,8 @@
 from django.core.urlresolvers import resolve
 from django.shortcuts import redirect
 
+from .actions import CustomerProxy
 from .conf import settings
-from .models import Customer
 
 
 class ActiveSubscriptionMiddleware(object):
@@ -12,9 +12,10 @@ class ActiveSubscriptionMiddleware(object):
             url_name = resolve(request.path).url_name
             if url_name not in settings.PINAX_STRIPE_SUBSCRIPTION_REQUIRED_EXCEPTION_URLS:
                 try:
-                    if not request.user.customer.has_active_subscription():
+                    customer = CustomerProxy.get_for_user(request.user)
+                    if not customer.has_active_subscription():
                         return redirect(
                             settings.PINAX_STRIPE_SUBSCRIPTION_REQUIRED_REDIRECT
                         )
-                except Customer.DoesNotExist:
+                except CustomerProxy.DoesNotExist:
                     return redirect(settings.PINAX_STRIPE_SUBSCRIPTION_REQUIRED_REDIRECT)
