+++ pinax/stripe/tests/test_managers.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from django.contrib.auth import get_user_model
 
 from . import TRANSFER_CREATED_TEST_DATA, TRANSFER_CREATED_TEST_DATA2
-from ..models import Event, Transfer, Customer, CurrentSubscription, Charge
+from ..actions import EventProxy, TransferProxy, CustomerProxy, CurrentSubscriptionProxy, ChargeProxy
 
 
 class CustomerManagerTest(TestCase):
@@ -19,14 +19,14 @@ class CustomerManagerTest(TestCase):
         period_end = datetime.datetime(2013, 4, 30, tzinfo=timezone.utc)
         start = datetime.datetime(2013, 1, 1, tzinfo=timezone.utc)
         for i in range(10):
-            customer = Customer.objects.create(
+            customer = CustomerProxy.objects.create(
                 user=User.objects.create_user(username="patrick{0}".format(i)),
                 stripe_id="cus_xxxxxxxxxxxxxx{0}".format(i),
                 card_fingerprint="YYYYYYYY",
                 card_last_4="2342",
                 card_kind="Visa"
             )
-            CurrentSubscription.objects.create(
+            CurrentSubscriptionProxy.objects.create(
                 customer=customer,
                 plan="test",
                 current_period_start=period_start,
@@ -36,14 +36,14 @@ class CustomerManagerTest(TestCase):
                 start=start,
                 quantity=1
             )
-        customer = Customer.objects.create(
+        customer = CustomerProxy.objects.create(
             user=User.objects.create_user(username="patrick{0}".format(11)),
             stripe_id="cus_xxxxxxxxxxxxxx{0}".format(11),
             card_fingerprint="YYYYYYYY",
             card_last_4="2342",
             card_kind="Visa"
         )
-        CurrentSubscription.objects.create(
+        CurrentSubscriptionProxy.objects.create(
             customer=customer,
             plan="test",
             current_period_start=period_start,
@@ -54,14 +54,14 @@ class CustomerManagerTest(TestCase):
             start=start,
             quantity=1
         )
-        customer = Customer.objects.create(
+        customer = CustomerProxy.objects.create(
             user=User.objects.create_user(username="patrick{0}".format(12)),
             stripe_id="cus_xxxxxxxxxxxxxx{0}".format(12),
             card_fingerprint="YYYYYYYY",
             card_last_4="2342",
             card_kind="Visa"
         )
-        CurrentSubscription.objects.create(
+        CurrentSubscriptionProxy.objects.create(
             customer=customer,
             plan="test-2",
             current_period_start=period_start,
@@ -74,50 +74,50 @@ class CustomerManagerTest(TestCase):
 
     def test_started_during_no_records(self):
         self.assertEquals(
-            Customer.objects.started_during(2013, 4).count(),
+            CustomerProxy.objects.started_during(2013, 4).count(),
             0
         )
 
     def test_started_during_has_records(self):
         self.assertEquals(
-            Customer.objects.started_during(2013, 1).count(),
+            CustomerProxy.objects.started_during(2013, 1).count(),
             12
         )
 
     def test_canceled_during(self):
         self.assertEquals(
-            Customer.objects.canceled_during(2013, 4).count(),
+            CustomerProxy.objects.canceled_during(2013, 4).count(),
             1
         )
 
     def test_canceled_all(self):
         self.assertEquals(
-            Customer.objects.canceled().count(),
+            CustomerProxy.objects.canceled().count(),
             1
         )
 
     def test_active_all(self):
         self.assertEquals(
-            Customer.objects.active().count(),
+            CustomerProxy.objects.active().count(),
             11
         )
 
     def test_started_plan_summary(self):
-        for plan in Customer.objects.started_plan_summary_for(2013, 1):
+        for plan in CustomerProxy.objects.started_plan_summary_for(2013, 1):
             if plan["current_subscription__plan"] == "test":
                 self.assertEquals(plan["count"], 11)
             if plan["current_subscription__plan"] == "test-2":
                 self.assertEquals(plan["count"], 1)
 
     def test_active_plan_summary(self):
-        for plan in Customer.objects.active_plan_summary():
+        for plan in CustomerProxy.objects.active_plan_summary():
             if plan["current_subscription__plan"] == "test":
                 self.assertEquals(plan["count"], 10)
             if plan["current_subscription__plan"] == "test-2":
                 self.assertEquals(plan["count"], 1)
 
     def test_canceled_plan_summary(self):
-        for plan in Customer.objects.canceled_plan_summary_for(2013, 1):
+        for plan in CustomerProxy.objects.canceled_plan_summary_for(2013, 1):
             if plan["current_subscription__plan"] == "test":
                 self.assertEquals(plan["count"], 1)
             if plan["current_subscription__plan"] == "test-2":
@@ -125,7 +125,7 @@ class CustomerManagerTest(TestCase):
 
     def test_churn(self):
         self.assertEquals(
-            Customer.objects.churn(),
+            CustomerProxy.objects.churn(),
             decimal.Decimal("1") / decimal.Decimal("11")
         )
 
@@ -133,7 +133,7 @@ class CustomerManagerTest(TestCase):
 class TransferManagerTest(TestCase):
 
     def test_transfer_summary(self):
-        event = Event.objects.create(
+        event = EventProxy.objects.create(
             stripe_id=TRANSFER_CREATED_TEST_DATA["id"],
             kind="transfer.created",
             livemode=True,
@@ -142,7 +142,7 @@ class TransferManagerTest(TestCase):
             valid=True
         )
         event.process()
-        event = Event.objects.create(
+        event = EventProxy.objects.create(
             stripe_id=TRANSFER_CREATED_TEST_DATA2["id"],
             kind="transfer.created",
             livemode=True,
@@ -151,8 +151,8 @@ class TransferManagerTest(TestCase):
             valid=True
         )
         event.process()
-        self.assertEquals(Transfer.objects.during(2012, 9).count(), 2)
-        totals = Transfer.objects.paid_totals_for(2012, 9)
+        self.assertEquals(TransferProxy.during(2012, 9).count(), 2)
+        totals = TransferProxy.paid_totals_for(2012, 9)
         self.assertEquals(
             totals["total_amount"], decimal.Decimal("19.10")
         )
@@ -176,11 +176,11 @@ class TransferManagerTest(TestCase):
 class ChargeManagerTests(TestCase):
 
     def setUp(self):
-        customer = Customer.objects.create(
+        customer = CustomerProxy.objects.create(
             user=get_user_model().objects.create_user(username="patrick"),
             stripe_id="cus_xxxxxxxxxxxxxx"
         )
-        Charge.objects.create(
+        ChargeProxy.objects.create(
             stripe_id="ch_1",
             customer=customer,
             charge_created=datetime.datetime(2013, 1, 1, tzinfo=timezone.utc),
@@ -189,7 +189,7 @@ class ChargeManagerTests(TestCase):
             fee=decimal.Decimal("3.42"),
             amount_refunded=decimal.Decimal("0")
         )
-        Charge.objects.create(
+        ChargeProxy.objects.create(
             stripe_id="ch_2",
             customer=customer,
             charge_created=datetime.datetime(2013, 1, 1, tzinfo=timezone.utc),
@@ -198,7 +198,7 @@ class ChargeManagerTests(TestCase):
             fee=decimal.Decimal("3.42"),
             amount_refunded=decimal.Decimal("10")
         )
-        Charge.objects.create(
+        ChargeProxy.objects.create(
             stripe_id="ch_3",
             customer=customer,
             charge_created=datetime.datetime(2013, 1, 1, tzinfo=timezone.utc),
@@ -207,7 +207,7 @@ class ChargeManagerTests(TestCase):
             fee=decimal.Decimal("3.42"),
             amount_refunded=decimal.Decimal("0")
         )
-        Charge.objects.create(
+        ChargeProxy.objects.create(
             stripe_id="ch_4",
             customer=customer,
             charge_created=datetime.datetime(2013, 4, 1, tzinfo=timezone.utc),
@@ -218,23 +218,23 @@ class ChargeManagerTests(TestCase):
         )
 
     def test_charges_during(self):
-        charges = Charge.objects.during(2013, 1)
+        charges = ChargeProxy.objects.during(2013, 1)
         self.assertEqual(charges.count(), 3)
 
     def test_paid_totals_for_jan(self):
-        totals = Charge.objects.paid_totals_for(2013, 1)
+        totals = ChargeProxy.objects.paid_totals_for(2013, 1)
         self.assertEqual(totals["total_amount"], decimal.Decimal("200"))
         self.assertEqual(totals["total_fee"], decimal.Decimal("6.84"))
         self.assertEqual(totals["total_refunded"], decimal.Decimal("10"))
 
     def test_paid_totals_for_apr(self):
-        totals = Charge.objects.paid_totals_for(2013, 4)
+        totals = ChargeProxy.objects.paid_totals_for(2013, 4)
         self.assertEqual(totals["total_amount"], decimal.Decimal("500"))
         self.assertEqual(totals["total_fee"], decimal.Decimal("6.04"))
         self.assertEqual(totals["total_refunded"], decimal.Decimal("15.42"))
 
     def test_paid_totals_for_dec(self):
-        totals = Charge.objects.paid_totals_for(2013, 12)
+        totals = ChargeProxy.objects.paid_totals_for(2013, 12)
         self.assertEqual(totals["total_amount"], None)
         self.assertEqual(totals["total_fee"], None)
         self.assertEqual(totals["total_refunded"], None)
