+++ pinax/stripe/tests/test_webhooks.py
@@ -9,7 +9,7 @@ from django.test.client import Client
 from mock import patch
 
 from . import TRANSFER_CREATED_TEST_DATA, TRANSFER_PENDING_TEST_DATA
-from ..models import Event, Transfer, EventProcessingException
+from ..actions import EventProxy, TransferProxy, EventProcessingExceptionProxy
 
 
 class TestWebhook(TestCase):
@@ -69,11 +69,11 @@ class TestWebhook(TestCase):
             content_type="application/json"
         )
         self.assertEquals(resp.status_code, 200)
-        self.assertTrue(Event.objects.filter(kind="transfer.created").exists())
+        self.assertTrue(EventProxy.objects.filter(kind="transfer.created").exists())
 
     def test_webhook_duplicate_event(self):
         data = {"id": 123}
-        Event.objects.create(stripe_id=123, livemode=True)
+        EventProxy.objects.create(stripe_id=123, livemode=True)
         msg = json.dumps(data)
         resp = Client().post(
             reverse("pinax_stripe_webhook"),
@@ -81,13 +81,13 @@ class TestWebhook(TestCase):
             content_type="application/json"
         )
         self.assertEquals(resp.status_code, 200)
-        self.assertTrue(EventProcessingException.objects.filter(message="Duplicate event record").exists())
+        self.assertTrue(EventProcessingExceptionProxy.objects.filter(message="Duplicate event record").exists())
 
 
 class TestTransferWebhooks(TestCase):
 
     def test_transfer_created(self):
-        event = Event.objects.create(
+        event = EventProxy.objects.create(
             stripe_id=TRANSFER_CREATED_TEST_DATA["id"],
             kind="transfer.created",
             livemode=True,
@@ -96,12 +96,12 @@ class TestTransferWebhooks(TestCase):
             valid=True
         )
         event.process()
-        transfer = Transfer.objects.get(stripe_id="tr_XXXXXXXXXXXX")
+        transfer = TransferProxy.objects.get(stripe_id="tr_XXXXXXXXXXXX")
         self.assertEquals(transfer.amount, decimal.Decimal("4.55"))
         self.assertEquals(transfer.status, "paid")
 
     def test_transfer_pending_create(self):
-        event = Event.objects.create(
+        event = EventProxy.objects.create(
             stripe_id=TRANSFER_PENDING_TEST_DATA["id"],
             kind="transfer.created",
             livemode=True,
@@ -110,12 +110,12 @@ class TestTransferWebhooks(TestCase):
             valid=True
         )
         event.process()
-        transfer = Transfer.objects.get(stripe_id="tr_adlkj2l3kj23")
+        transfer = TransferProxy.objects.get(stripe_id="tr_adlkj2l3kj23")
         self.assertEquals(transfer.amount, decimal.Decimal("9.41"))
         self.assertEquals(transfer.status, "pending")
 
     def test_transfer_paid_updates_existing_record(self):
-        event = Event.objects.create(
+        event = EventProxy.objects.create(
             stripe_id=TRANSFER_CREATED_TEST_DATA["id"],
             kind="transfer.created",
             livemode=True,
@@ -202,7 +202,7 @@ class TestTransferWebhooks(TestCase):
             "pending_webhooks": 1,
             "type": "transfer.paid"
         }
-        paid_event = Event.objects.create(
+        paid_event = EventProxy.objects.create(
             stripe_id=data["id"],
             kind="transfer.paid",
             livemode=True,
@@ -211,5 +211,5 @@ class TestTransferWebhooks(TestCase):
             valid=True
         )
         paid_event.process()
-        transfer = Transfer.objects.get(stripe_id="tr_XXXXXXXXXXXX")
+        transfer = TransferProxy.objects.get(stripe_id="tr_XXXXXXXXXXXX")
         self.assertEquals(transfer.status, "paid")
